requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
pydentic